requirements.txt
openai
google-generativeai
pinecone
beautifulsoup4
requests
python-dotenv
